check dest dir

+++ unreal/Puerts/Source/JsEnv/JsEnv.Build.cs
@@ -181,13 +181,16 @@ public class JsEnv : ModuleRules
             "puerts/index.d.ts",
             "ffi/index.d.ts"
         };
-        foreach (var Source in SourceFiles)
+        destDirName = Path.GetFullPath(Path.Combine(ModuleDirectory, "..", "..", "..", "..", "Typing"));
+        if(Directory.Exists(destDirName))
         {
-            string src = Path.GetFullPath(Path.Combine(ModuleDirectory, "..", "..", "Typing", Source));
-            string dest =
-                Path.GetFullPath(Path.Combine(ModuleDirectory, "..", "..", "..", "..", "Typing", Source));
-            // Logger.LogDebug("Copy " + Source + " from " + src + " to " + dest);
-            File.Copy(src, dest, true);
+            foreach (var Source in Sources)
+            {
+                string src = Path.GetFullPath(Path.Combine(ModuleDirectory, "..", "..", "Typing", Source));
+                string dest = Path.GetFullPath(Path.Combine(destDirName, Source));
+                // Logger.LogDebug("Copy " + Source + " from " + src + " to " + dest);
+                File.Copy(src, dest, true);
+            }
         }
     }
